Move filesystem operations into a seperate header to make it easier for contributors who don't have a calculator or emulator.
