+++ src/main.c
@@ -1,6 +1,7 @@
 #include "graphics.h"
 #include "button_graphics.h"
 #include "key_handler.h"
+#include "filesystem.h"
 #include <math.h>
 
 
@@ -48,38 +49,6 @@ typedef struct {
 player_t player;
 
 const char *appvarName = "slota";
-
-void SaveData(void) {
-    ti_var_t slota;
-    ti_CloseAll();
-    if (slota = ti_Open("AMGDAT","w+")) {
-        ti_Write(&player,sizeof(player),1,slota);
-        //ti_Write(&game,sizeof(game),1,slota);
-        //ti_Write(&dummies,sizeof(dummies),1,slota);
-    }
-    ti_SetArchiveStatus(true, slota);
-}
-
-void LoadData(void) {
-    ti_var_t slota;
-    ti_CloseAll();
-    if (slota = ti_Open("AMGDAT", "r")) {
-        ti_Read(&player, sizeof(player),1,slota);
-        //ti_Read(&game, sizeof(game),1,slota);
-        //ti_Read(&dummies, sizeof(dummies),1,slota);
-    }
-}
-
-void CreateSave(void) {
-    ti_var_t slota;
-    ti_CloseAll();
-    if (slota = ti_Open("AMGDAT","w+")) {
-        ti_Write(&player, sizeof(player), 1, slota);
-        ti_SetArchiveStatus(true, slota);
-        //ti_Write(&dummies, sizeof(dummies),1,slota);
-        //ti_Write(&game, sizeof(game), 1, slota);
-    }
-}
 
 void doCollisions() {
     //base it off of rooms and player direction
